requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.4.0
python-dotenv==1.0.1 
ctransformers
sqlalchemy==2.0.25
aiosqlite==0.19.0 
bcrypt==4.0.1
passlib==1.7.4